Move the choice-class to another directory
